# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Server API",
        "description": "API сайта",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://siveno.shop"
        }
    ],
    "paths": {
...
        "/api/russian-post/create-shipment": {
            "post": {
                "tags": [
                    "RussianPost"
                ],
                "summary": "Создание отправления через Почту России",
                "description": "Создает отправление (заказ) через Почту России с параметрами получателя и посылки. Использует API PUT /1.0/user/backlog.",
                "operationId": "russianPostCreateShipment",
                "requestBody": {
                    "description": "Данные отправления",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_num",
                                    "recipient",
                                    "package"
                                ],
                                "properties": {
                                    "order_num": {
                                        "description": "Уникальный номер заказа",
                                        "type": "string",
                                        "example": "ORDER-12345"
                                    },
                                    "recipient": {
                                        "description": "Данные получателя",
                                        "required": [
                                            "name",
                                            "postcode",
                                            "region",
                                            "city",
                                            "street",
                                            "house",
                                            "phone"
                                        ],
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Иван Иванов"
                                            },
                                            "postcode": {
                                                "type": "string",
                                                "example": "190000"
                                            },
                                            "region": {
                                                "type": "string",
                                                "example": "Ленинградская обл."
                                            },
                                            "city": {
                                                "type": "string",
                                                "example": "Санкт-Петербург"
                                            },
                                            "street": {
                                                "type": "string",
                                                "example": "ул. Пушкина"
                                            },
                                            "house": {
                                                "type": "string",
                                                "example": "10"
                                            },
                                            "room": {
                                                "type": "string",
                                                "example": "5"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "79991234567"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "package": {
                                        "description": "Параметры посылки",
                                        "required": [
                                            "weight",
                                            "length",
                                            "width",
                                            "height"
                                        ],
                                        "properties": {
                                            "weight": {
                                                "description": "Вес в кг",
                                                "type": "number",
                                                "example": 1.5
                                            },
                                            "length": {
                                                "description": "Длина в см",
                                                "type": "integer",
                                                "example": 30
                                            },
                                            "width": {
                                                "description": "Ширина в см",
                                                "type": "integer",
                                                "example": 20
                                            },
                                            "height": {
                                                "description": "Высота в см",
                                                "type": "integer",
                                                "example": 10
                                            },
                                            "value": {
                                                "description": "Объявленная ценность в копейках",
                                                "type": "integer",
                                                "example": 5000
                                            },
                                            "fragile": {
                                                "description": "Хрупкий груз",
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Отправление успешно создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result-ids": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer",
                                                "example": 123456
                                            }
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "position": {
                                                        "type": "integer",
                                                        "example": 0
                                                    },
                                                    "error-codes": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "code": {
                                                                    "type": "string",
                                                                    "example": "UNDEFINED"
                                                                },
                                                                "description": {
                                                                    "type": "string",
                                                                    "example": "Ошибка обработки запроса"
                                                                },
                                                                "details": {
                                                                    "type": "string",
                                                                    "example": "index-to is invalid"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка создания отправления через Почту России",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка создания отправления через Почту России"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/russian-post/track-shipment": {
            "post": {
                "tags": [
                    "RussianPost"
                ],
                "summary": "Отслеживание отправления через Почту России",
                "description": "Возвращает информацию о статусе отправления по номеру отслеживания.",
                "operationId": "russianPostTrackShipment",
                "requestBody": {
                    "description": "Номер отслеживания отправления",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "trackingNumber"
                                ],
                                "properties": {
                                    "trackingNumber": {
                                        "description": "Номер отслеживания отправления",
                                        "type": "string",
                                        "example": "PO123456789RU"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Информация об отправлении получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка отслеживания отправления через Почту России",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка отслеживания отправления через Почту России"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/russian-post/search-order": {
            "get": {
                "tags": [
                    "RussianPost"
                ],
                "summary": "Поиск отправления по номеру заказа",
                "description": "Возвращает информацию об отправлении по order-num (идентификатору заказа магазина).",
                "operationId": "russianPostSearchOrder",
                "parameters": [
                    {
                        "name": "orderNumber",
                        "in": "query",
                        "description": "Номер заказа (order-num)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об отправлении найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Заказ не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Заказ не найден"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка поиска заказа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка поиска отправления через Почту России"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите токен в формате Bearer {your_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}